Finally implemented custom item textures
Oh my god this was a nightmare.
I ended up having to abandon my original spritesheet plan and instead rely on predicates, but even just implementing that was horrible.
Also, there is now a limit of 256 for each ingredient type due to limitations with the current system (4 bytes for each int is combined and converted to a float, because predicates are floats).
I also had to set up my own UnclampedModelPredicateProvider class, because the "Unclamped" class clamps the output.
I will also probably have to set up additional code to allow me to "reserve" ingredient slots, because if I don't do this I will be unable to remove ingredients without ruining the textures of all the ingredients below in the spreadsheet.
Fortunately I shouldn't need to do something similar to this with entity textures.